Remove options menu from WebmakerActivity and add to Map activity directly
